fix(version): fold the host-id heal + tor identity migration into beta.10
beta.10 is still unreleased, so the migration ships as part of it rather
than a new beta.11 — drop the version bump and move up()/post_up() logic
into v0_4_0_beta_10. Note this means dev boxes whose db already records
beta.10 will not re-run the migration (their stored version equals
current); only boxes upgrading from <= beta.9 get the heal + tor
re-pointing.
