+++ projects/start-sdk/lib/StartSdk.ts
@@ -72,7 +72,7 @@ import { createVolumes } from './util/Volume'
 import { getDataVersion, setDataVersion } from './version'
 
 /** The minimum StartOS version required by this SDK release */
-export const OSVersion = testTypeVersion('0.4.0-beta.11')
+export const OSVersion = testTypeVersion('0.4.0-beta.10')
 
 // prettier-ignore
 type AnyNeverCond<T extends any[], Then, Else> =
