+++ shared-libs/crates/start-core/Cargo.toml
@@ -15,7 +15,7 @@ license = "MIT"
 name = "start-core"
 readme = "README.md"
 repository = "https://github.com/Start9Labs/start-technologies"
-version = "0.4.0-beta.11" # VERSION_BUMP
+version = "0.4.0-beta.10" # VERSION_BUMP
 
 [lib]
 name = "start_core"
