+++ shared-libs/crates/start-core/src/version/mod.rs
@@ -66,7 +66,6 @@ mod v0_4_0_alpha_23;
 mod v0_4_0_beta_0;
 mod v0_4_0_beta_1;
 mod v0_4_0_beta_10;
-mod v0_4_0_beta_11;
 mod v0_4_0_beta_2;
 mod v0_4_0_beta_3;
 mod v0_4_0_beta_4;
@@ -76,7 +75,7 @@ mod v0_4_0_beta_7;
 mod v0_4_0_beta_8;
 mod v0_4_0_beta_9;
 
-pub type Current = v0_4_0_beta_11::Version; // VERSION_BUMP
+pub type Current = v0_4_0_beta_10::Version; // VERSION_BUMP
 
 impl Current {
     #[instrument(skip(self, db))]
@@ -218,8 +217,7 @@ enum Version {
     V0_4_0_beta_7(Wrapper<v0_4_0_beta_7::Version>),
     V0_4_0_beta_8(Wrapper<v0_4_0_beta_8::Version>),
     V0_4_0_beta_9(Wrapper<v0_4_0_beta_9::Version>),
-    V0_4_0_beta_10(Wrapper<v0_4_0_beta_10::Version>),
-    V0_4_0_beta_11(Wrapper<v0_4_0_beta_11::Version>), // VERSION_BUMP
+    V0_4_0_beta_10(Wrapper<v0_4_0_beta_10::Version>), // VERSION_BUMP
     Other(exver::Version),
 }
 
@@ -296,8 +294,7 @@ impl Version {
             Self::V0_4_0_beta_7(v) => DynVersion(Box::new(v.0)),
             Self::V0_4_0_beta_8(v) => DynVersion(Box::new(v.0)),
             Self::V0_4_0_beta_9(v) => DynVersion(Box::new(v.0)),
-            Self::V0_4_0_beta_10(v) => DynVersion(Box::new(v.0)),
-            Self::V0_4_0_beta_11(v) => DynVersion(Box::new(v.0)), // VERSION_BUMP
+            Self::V0_4_0_beta_10(v) => DynVersion(Box::new(v.0)), // VERSION_BUMP
             Self::Other(v) => {
                 return Err(Error::new(
                     eyre!("unknown version {v}"),
@@ -366,8 +363,7 @@ impl Version {
             Version::V0_4_0_beta_7(Wrapper(x)) => x.semver(),
             Version::V0_4_0_beta_8(Wrapper(x)) => x.semver(),
             Version::V0_4_0_beta_9(Wrapper(x)) => x.semver(),
-            Version::V0_4_0_beta_10(Wrapper(x)) => x.semver(),
-            Version::V0_4_0_beta_11(Wrapper(x)) => x.semver(), // VERSION_BUMP
+            Version::V0_4_0_beta_10(Wrapper(x)) => x.semver(), // VERSION_BUMP
             Version::Other(x) => x.clone(),
         }
     }
